# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口详情",
        "title": "gin-restful-api模板",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user": {
            "get": {
                "description": "查询用户列表",
                "tags": [
                    "Api.User"
                ],
                "summary": "查询用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "名字",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "年龄",
                        "name": "age",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "性别",
                        "name": "gender",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    },
                    "1": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    }
                }
            },
            "post": {
                "description": "新增用户",
                "tags": [
                    "Api.User"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entuser.CreateBody"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    },
                    "1": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "description": "查询用户详情",
                "tags": [
                    "Api.User"
                ],
                "summary": "查询用户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    },
                    "1": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    }
                }
            },
            "put": {
                "description": "更新用户",
                "tags": [
                    "Api.User"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entuser.UpdateBody"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    },
                    "1": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/api.ResponseJson"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除用户",
                "tags": [
                    "Api.User"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "请求成功",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "1": {
                        "description": "请求失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ResponseJson": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "entuser.CreateBody": {
            "type": "object",
            "required": [
                "age",
                "gender",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 200
                },
                "createdAt": {
                    "type": "string"
                },
                "gender": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entuser.UpdateBody": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer",
                    "maximum": 200
                },
                "gender": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}